requirements.txt
click==8.0.1
mysql-connector==2.2.9
numpy==1.21.0
opencv-python==4.5.2.54
pandas==1.2.5
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.1.0
six==1.16.0
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
